requirements.txt
auto_mix_prep==0.2.0
dtaidistance==2.3.10
einops==0.6.1
extensions==0.4
fastdtw==0.3.4
fitter==1.5.2
matplotlib==3.7.2
numpy==1.24.4
ont_fast5_api==4.1.1
opt_einsum==3.3.0
pod5==0.2.4
pykeops==2.1.2
pyslow5==1.0.0
python_Levenshtein==0.21.1
pytorch_lightning==2.0.4
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.11.1
tensorflow==2.13.0
torch==2.0.1
tqdm==4.65.0
wandb==0.15.4
